Move test together
