+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "lint": "standard | snazzy",
-    "unit": "mocha */test/*.test.js",
+    "unit": "mocha test/*.test.js",
     "test": "npm run lint && npm run unit",
     "start": "node index.js"
   },
